Add files via upload
Only the forest plot dataset (data.figure2.xlsx" and "data.figure3.xlsx") is provided for testing the codes. The raw data that support the findings of this study are not publicly available for confidentiality reasons, as these patients may be re-identified through various techniques, such as data linkage or combining datasets. The processed data are available on reasonable request to the corresponding author, LBY, with each request subject to ethical and legislative review from the respective data sources.